Implementada realtime database para controlar la presencia de los usuarios en una sala

+++ src/app/pages/bingo/bingo/bingo.ts
@@ -29,30 +29,35 @@ export class BingoComponent {
   roomId!: string | null
   private routeSub!: Subscription;
   ngOnInit() {
-    this.routeSub = this.route.paramMap.subscribe(params => {
-    this.roomId  = params.get('id');
-    this.isAdmin = this.router.url.includes('/admin/');
-    if (!this.roomId) {
-      this.router.navigate(['/lobby']);
-      return;
-    }
-    if (this.isAdmin) {
-      this.bingoService.createRoom(this.roomId);
-    } else {
-      this.bingoService.joinRoom(this.roomId);
-    }
-  });
-}
+    this.routeSub = this.route.paramMap.subscribe(async params => {
+      this.roomId = params.get('id');
+      this.isAdmin = this.router.url.includes('/admin/');
+      if (!this.roomId) {
+        this.router.navigate(['/lobby']);
+        return;
+      }
+      if (this.isAdmin) {
+        await this.bingoService.createRoom(this.roomId);
+      } else {
+        await this.bingoService.joinRoom(this.roomId);
+      }
+    });
+  }
 
   selectTab(tab: 'tablero' | 'orden' | 'players') {
     this.selectedTab = tab;
   }
 
-  ngOnDestroy(): void {
+  async ngOnDestroy(): Promise<void> {
     this.routeSub.unsubscribe();
+    this.bingoService.unsubscribe();
     const user = this.authService.getCurrentUser();
-    if (user) {
-      this.bingoService.removePlayerFromRoom(user.uid);
+    if (user && this.roomId) {
+      // 1. Limpia presencia en Realtime DB
+      await this.bingoService.clearUserPresence(this.roomId, user.uid);
+
+      // 2. (Opcional) elimina al jugador también de Firestore
+      await this.bingoService.removePlayerFromRoom(user.uid);
     }
   }
